Add stake delegation and distribution to validator table
